+++ src/main/resources/release_notes.md
@@ -38,6 +38,7 @@
 
 * Fix empty list of upcoming games in Lineup panel after first download.
 * Added support of all 13 **Player Categories** (#2418)
+* Fix scroll behavior of lineup rating panel (#2365)
 
 ### Training